+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 -   Dropped Python 3.7 and 3.8 support.
 -   Added support for Python 3.12.
+-   New parameters have been added to the `Variation` class. The old syntax 
+    is still supported but has been deprecated.
 
 ## [0.3.3](https://github.com/dldevinc/variations/tree/v0.3.3) - 2023-04-25